sadfddas

+++ gdx-game/src/br/letonai/jewels/MyGdxGame.java
@@ -36,8 +36,8 @@ public class MyGdxGame implements ApplicationListener
 	{        
 		Vector3 touchPos = new Vector3();
 		Vector2 gravity = new Vector2(0,-1);
-		double t=TimeUtils.millis();
-	
+		float t=(float) TimeUtils.millis();
+		
 		if(false && Gdx.input.isTouched()){
 			
 			touchPos.set(Gdx.input.getX(),Gdx.input.getY(),0);
@@ -56,7 +56,7 @@ public class MyGdxGame implements ApplicationListener
 		
 		player.setX(touchPos.x-player.getTexture().getWidth()/2);
 		//player.setY((Gdx.graphics.getHeight()-touchPos.y)-player.getTexture().getHeight()/2);
-		player.setY((Gdx.graphics.getHeight()-MathUtils.sin((float) t))-player.getTexture().getHeight()/2);
+		player.setY((Gdx.graphics.getHeight()-MathUtils.sin(t))-player.getTexture().getHeight()/2);
 		batch.end();
 	}
